requirements_eval.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==2.14.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
faiss-cpu==1.12.0
fastjsonschema==2.21.2
filelock==3.20.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.4.0
google-api-core==2.26.0
google-auth==2.41.1
google-generativeai==0.4.1
googleapis-common-protos==1.70.0
groq==0.32.0
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
instructor==1.11.3
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.42
lark==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
mypy_extensions==1.1.0
narwhals==2.9.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.6.0
openai==1.109.1
orjson==3.11.3
ormsgpack==1.11.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.5
pypdf==6.1.1
pypdfium2==4.30.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scikit-network==0.33.3
scipy==1.16.2
Send2Trash==1.8.3
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.54
stack-data==0.6.3
streamlit==1.42.2
sympy==1.14.0
tenacity==8.2.3
terminado==0.18.1
text-generation==0.7.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.19.2
types-python-dateutil==2.9.0.20251008
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.14
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
